022319 code maintenance
- Renamed Sim Tactics to SimTactics Classic in order to focus on the new partial rebuild based on UrhoSharp, SimTactics. Little confusing, I know.
- Removed FSO.Server.Core.Api for the reason stated above
- Updated the Bootstrap dependency in FSO.Server.Api from 3.0.0 to 3.4.1 due to a potential security vulnerability, CVE-2019-8331. If there is ANY problems, please let me know. Because I have no idea where this is used.
- Ported FSO.Server.Watchdog, the TeamCity-based updater, over to .NET Core. I will also be migrating it's config file from Ini to Toml format since I really don't think it's necessary to use the former everywhere just because TSO did. They're much better solutions that don't involve always rewriting the parser.

+++ Src/tso.simantics/primitives/VMPushInteraction.cs
@@ -5,13 +5,8 @@
  */
 
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using FSO.SimAntics.Engine;
 using FSO.Files.Utils;
 using FSO.Files.Formats.IFF.Chunks;
-using FSO.Content;
 using System.IO;
 
 namespace FSO.SimAntics.Engine.Primitives
@@ -25,8 +20,8 @@ namespace FSO.SimAntics.Engine.Primitives
             var operand = (VMPushInteractionOperand)args;
             VMEntity interactionSource;
 
-            if (operand.ObjectInLocal) interactionSource = context.VM.GetObjectById((short)context.Locals[operand.ObjectLocation]);
-            else interactionSource = context.VM.GetObjectById((short)context.Args[operand.ObjectLocation]);
+            if (operand.ObjectInLocal) interactionSource = context.VM.GetObjectById(context.Locals[operand.ObjectLocation]);
+            else interactionSource = context.VM.GetObjectById(context.Args[operand.ObjectLocation]);
 
             short priority = 0;
             VMQueueMode mode = VMQueueMode.Normal;
@@ -52,7 +47,7 @@ namespace FSO.SimAntics.Engine.Primitives
 
             var action = interactionSource.GetAction(operand.Interaction, context.StackObject, context.VM.Context, false);
             if (action == null) return VMPrimitiveExitCode.GOTO_FALSE;
-            if (operand.UseCustomIcon) action.IconOwner = context.VM.GetObjectById((short)context.Locals[operand.IconLocation]);
+            if (operand.UseCustomIcon) action.IconOwner = context.VM.GetObjectById(context.Locals[operand.IconLocation]);
             action.Mode = mode;
             action.Priority = priority;
             action.Flags |= TTABFlags.FSOSkipPermissions;
